include the ACH/CAH logic in the complications code (402)
In the current version of the code, the logic to only include complications in short term acute care hospitals and CAHs is performed in the 300 code rather than here. When I cleaned up the flags in 300 to directly use the source files from the 402 output, it skipped over the hospital criteria so it was counting complications at IRFs, which should be excluded. The change here is to apply the hospital criteria using the elig_hosp flag so that the output files from the 402 code are cleaner.